requirements.txt
asgiref==3.9.1
diff-match-patch==20241021
Django==5.2.5
django-import-export==4.3.9
django-widget-tweaks==1.5.0
pillow==11.3.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
